feat 교환일기 페이지(아직 교환일기시작안했을때) 추가
